+++ lib/src/server/server.dart
@@ -4,10 +4,10 @@
 
 library application;
 
+import 'dart:async';
 import 'dart:convert' show UTF8;
 import 'dart:io';
 
-import '../app_engine_request_handler.dart';
 import 'context_registry.dart';
 import 'http_wrapper.dart';
 
@@ -26,13 +26,18 @@ class AppEngineHttpServer {
   final int _port;
 
   final String _localHostname;
+  final Completer _shutdownCompleter = new Completer();
+  int _pendingRequests = 0;
+
   HttpServer _httpServer;
 
   AppEngineHttpServer(this._contextRegistry,
                       {String hostname: '0.0.0.0', int port: 8080})
       : _localHostname = _getHostname(), _hostname = hostname, _port = port;
 
-  void run(AppEngineRequestHandler applicationHandler) {
+  Future get done => _shutdownCompleter.future;
+
+  void run(applicationHandler(request, context)) {
     var serviceHandlers = {
         '/_ah/start' : _start,
         '/_ah/health' : _health,
@@ -60,49 +65,58 @@ class AppEngineHttpServer {
           }
         }
 
-        _contextRegistry.add(appengineRequest).then((_) {
-          /*
-           * This sets the 'Server' header in the http response to the hostname
-           * of the machine the application is running on.
-           * It seems like the AppEngine VmRuntime stub (on the other end) will
-           * not accept the response if we use the default value.
-           */
-          appengineRequest.response.headers.set('Server', _localHostname);
-          appengineRequest.response.registerHook(
-              () => _contextRegistry.remove(appengineRequest));
+        _pendingRequests++;
+        var context = _contextRegistry.add(appengineRequest);
+        /*
+         * This sets the 'Server' header in the http response to the hostname
+         * of the machine the application is running on.
+         * It seems like the AppEngine VmRuntime stub (on the other end) will
+         * not accept the response if we use the default value.
+         */
+        appengineRequest.response.headers.set('Server', _localHostname);
+        appengineRequest.response.registerHook(
+            () => _contextRegistry.remove(appengineRequest));
 
-          appengineRequest.response.done.catchError((error) {
-            _info("Error while handling response: $error");
-          });
-
-          handler(appengineRequest);
+        appengineRequest.response.done.catchError((error) {
+          _info("Error while handling response: $error");
+          _pendingRequests--;
+          _checkShutdown();
         });
+
+        handler(appengineRequest, context);
       });
     });
   }
 
-  void _start(HttpRequest request) {
+  void _start(HttpRequest request, _) {
     request.drain().then((_) {
       _sendResponse(request.response, HttpStatus.OK, "ok");
     });
   }
 
-  void _health(HttpRequest request) {
+  void _health(HttpRequest request, _) {
     request.drain().then((_) {
       _sendResponse(request.response, HttpStatus.OK, "ok");
     });
   }
 
-  void _stop(HttpRequest request) {
+  void _stop(HttpRequest request, _) {
     request.drain().then((_) {
       if (_httpServer != null) {
-        _httpServer.close().then((_){
+        _httpServer.close().then((_) {
+          _httpServer = null;
           _sendResponse(request.response, HttpStatus.OK, "ok");
         });
       } else {
         _sendResponse(request.response, HttpStatus.CONFLICT, "fail");
       }
     });
+  }
+
+  _checkShutdown() {
+    if (_pendingRequests == 0 && _httpServer == null) {
+      _shutdownCompleter.complete();
+    }
   }
 
   void _sendResponse(HttpResponse response, int statusCode, String message) {